add_readme_training
