--- Research/training/command.md
@@ -1,9 +1,0 @@
-# training
-'''
-python train.py --train_path ../datasets/KDDTrain+.txt --save_path ../Output_model/NSL_KDD/random_forest_model.pkl
-'''
-
-# inference
-'''
-python inference.py --test_path path_to_csv_file_test --save_path ../Output_model/NSL_KDD/random_forest_model.pkl
-'''
